Move OrderValidator apart
